Move Cli#validate to a ValidateCommand class
